fix bugs: b2200 reset io error

+++ src/bledriver/util/gl_hal.h
@@ -30,7 +30,7 @@ typedef struct {
     uint32_t baudRate;
     uint32_t flowcontrol;
     uint8_t rst_trigger;
-    uint8_t rst_gpio;
+    uint16_t rst_gpio;
 }hw_cfg_t;
 
 /*Initilize the hal*/
